Data Mining/Stock Market Data till (15-10-2024)/requirements.txt
nselib==1.2
mysql-connector==2.2.9
pandas
numpy
DateTime
python-dateutil
tqdm
